Merging branch 'machine' (#3)
The machine is not 100% complete. However the main console is functional now and the physics button press is functional.
Note1: No need to add any [CameraRig] or OVRCameraRig, as VRTK picks the setup you are on and switches automatically.
Note2: VirtualGrasp hands are not added yet. They will be added asap.