Adding MainScene Under folder Assests/
This will be the main scene. I copied the scene I'm working on and pasted it in the Assets folder so it's easier for everybody to copy it to their folder when needed.
I will be working on this scene directly, so no modifications are allowed to this scene, only copying.