Feat/ui hint system (#61)
* UI with VG rotation problem

* organizing file structure

* setting ui_02

* wip

* wip - modifying animations

* Petter 21 ui hintsystem animations and refinements (#59)

* Animation for UI system

Adding elements and new animations together with Backer. Also the knife material

* Animations for Hintsystem_03

Animations like the power button activating the hint system when the player presses it for the first time.

* HintSystem is working (in place)

* HintSystem ready alternative 1

* done!

* Petter 21 ui hintsystem animations and refinements (#60)

* Animation for UI system

Adding elements and new animations together with Backer. Also the knife material

* Animations for Hintsystem_03

Animations like the power button activating the hint system when the player presses it for the first time.

* Balloon_UI animations

The sign that we can use for marking locations and buttons. Remember that the "destination point" is now a part of the prefab and should be attached to the button or destination you want to show. Also Rememeber that changing the position of the line render object isteslf will cause its points to move as well. THis can cause confusion so stay alert.